Create enhanced-shortcodes.php
As discussed in the comments section (http://www.itthinx.com/plugins/groups-gravityforms/), here's my take on the shortcode file.

As far as I know, all original functinality is intact.

This edit does a couple of things:

Listing groups the user is a member of: Add an option to display a "leave group" button for each group. Use leavebutton="yes" in the shortcode to display it, e.g like this: [groups_user_groups leavebutton="1"]

Listing all groups available: Add an option to display a "join group" button for each group. Use joinbutton="yes" in the shortcode to display it, e.g. like this: [groups_groups joinbutton="1"]

Listing all groups available: Add an option to not display groups the logged in user is already a member of. Use exclude_excisting="yes" in the shortcode to enable it, e.g. like this: [[groups_groups exclude_excisting="1"]

These last two are ment to be used together, as it would make little sense in asking the user to join a group she's already a member of.

I'm no PHP expert, so there's probably improvements to be made on the actual code. Feel free to look into it.

All the best,
Eivind