Update class-groups-shortcodes.php
So, a new update here. Still the same issues with not reusing codes by calling on the join_group() and leave_group() functions. Somehow, I couldn't get them to work, so, quite a lot of work around.

I also had problems leaving and joining several groups, and figured it had something to do with wp_nonce(). So, I added more detailed nonce's, and the problems are gone now.

There's also an issue with the actual joining and leaving. In order for this code to work, the join and leave shortcodes need to be included on the same page as the other shortcodes. Because these shortcodes are processed after the listings, there's also some logic involved in updating the lists to show/hide the recent join/leave. (But, come to think about it now, I guess I might have saved me that bit of work by putting the leave/join shortcodes before the list shortcodes on the WP page..).

Ah, well. This is mostly rubbish, but in the unlikely chance of someone needing this functionality before you get a chance to do this properly in the master branch, I guess it won't hurt to post the code as is.